requirements.txt
click==8.1.3
pydantic==1.9.2
PyPDF2==2.10.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.3.0
